Create separate subfolder for TP tests
